overhaul of infer_sequential_source: new data inclusion/exclusion criteria
Following the changes to infer_singledot_source I have changed the logic of
which data enters the analysis. Specifically, I now include for each trial as
many dots in the sequential analysis as there are available data. Also, I now
do not exclude whole trials, when some data or regressor values for one of the
desired dots is missing in that trial. Instead, I only exclude the dots for
the missing data.

other important changes:
- exclusion of data based on closeness to RT (toolate=-200)
- exclusion of timed out trials
- local normalisation of design matrix

As these changes mean that at each time point of the sequential analysis a
potentially varying number of data points enter the regression, I now save the
data count for each time point and subject as well as the condition numbers of
the design matrix. As in infer_singledot_source I also exclude subjects from
second-level analysis for which there were less than mindata (30) data points,
but the minimum number of data points in any regression across subjects and
time points was 137 so that no subject was excluded by this criterion.
Interestingly, the data counts drop approximately linear here. That is probably
because the RT has an effect starting from time point 0 and moving forward in
time just excludes more and more data. The mean data count for the committed
analysis starts at 4656 and drops to 1583 at time point 690. Condition numbers
of the design matrix mostly stayed below 5.5 and only went up considerably as
time from dot onset increased for subjects 24, 32 and 33, but also only to
values between 6 and 8.

Results
-------
are stored in
	source_sequential_201801261754.h5

and needed about 1 day to compute.

The plot of average effect magnitudes across the whole brain for different time
points shows that the time course of the average effect is intact with peaks
for dot_x at 120 and 180 ms. The peaks for 320 and 400 ms, however, are now 20
ms later. The plot also shows that the analysis becomes unreliable for time
points beyond 500 ms as the average effect magnitudes for the permuted data
rise and reach similar values as for the original data.

Compared to the previous results, the number of significant clusters after
FDR-correction has dropped considerably. Probably this is because the
exclusion of data close to the response. The surviving effects of dot-x in the
time window from 400 to 500 ms after dot onset are almost exclusively in
somatosensory and motor areas (including M1) and posterior cingulate. The first
peak at 120 ms contains almost only visual and posterior cingulate effects. The
second peak around 180 ms has the strongest (of its few) effects in parietal
areas. From 300 to 390 ms effects are in motor areas, posterior, mid and
anterior cingulate.

Apart from dot-x only accev has significant effects after FDR-correction.
Summed across clusters of the same area, the areas with the largest
accev-effects are L_4, L_3a, R_7m, R_V6, L_31pv, R_31pd, R_v23ab, R_FEF and
L_v23ab. So it's interesting that the largest effects are in motor areas in the
left hemisphere, but in visual-parietal areas in the right hemisphere, but this
may be an artefact of cluster selection.