normalisation of source data within time points and subjects
Previously I only normalised the source currents globally, i.e., across trials,
subjects and time points. I assumed that this would not influence results for,
e.g., dot-x correlations except in the scale of the estimated betas, because
the regression anyway uses a subject and time point specific intercept. For the
sequential analysis, however, the intercept reaches across several time points.
I, therefore, thought that it is possible that the average over trials changes
across time points and that this may obscure the correlations of interest. So I
thought about including time-point specific intercepts in the regression model,
but opted for the simpler normalisation within time points of single subjects
instead of global normalisation of the source data. The point is that a
changing mean signal across time points adds unnecessary variance to the data
when I concatenate data of several time points within a trial for the
sequential analysis. By normalising across trials, but within time points I
remove this variance and potentially improve inference for the underlying
correlations.

Results
-------
are stored in:
	source_sequential_201801291241.h5

I did not observe any meaningful improvements in inference results, i.e., I
neither clearly observed that significant effects became more reliable, nor did
I observe substantially new effects compared to the previous run with global
normalisation. For example, second-level t-values for dot-x at 420 ms after dot
onset differed only slightly across brain areas (most t-value differences were
within [-0.2, 0.2] with a mean of 0.003 and a std of 0.12) and there was no
clear relation between the size of the difference and the t-values. This was
similar for the other regressors except for the intercept which had a std of
t-value differences across brain areas of 1.52 and a clear anti-correlation
between the intercept t-value for within time point normalisation and the
t-value difference.

Easier interpretation of these results is with mlog10p which is an unsigned
measure or reliability. For dot-x at 420 ms mlog10p differences have a mean of
0.01 and std of 0.1 indicating slightly more reliable estimates with global
normalisation, but overall small changes in estimates. The intercept at the
same time point had mlog10p differences of 0.2 +- 0.84 indicating larger
effects of normalisation which should be expected. Of the other regressors only
percupt-x had somewhat similarly high differences with 0.2 +- 0.4.

Within a brain area and across time points instead of across brain areas the
results are similar. For example for L_4 the mlog10p-values with global
normalisation are distributed according to 1.69 +- 1.81 across time points
while the mlog10p-differences are in 0.08 +- 0.15 indicating that the effects
of different normalisation are small and slightly favour global normalisation
over normalisation within time points. For L-v23ab the normalisation effects
appear even smaller: 1.88 +- 1.48 vs. 0.01 +- 0.1. Searched across all brain
areas the largest mean mlog10p differences for dot-x occur for L_AIP (-0.17),
R_V2 (-0.11), R_IFSp (0.1), L_3a (0.09) and R_3a (-0.09). Again, these effects
are small and it's strange that within time point normalisation leads at the
same time to larger effects in R_3a and smaller effects in L_3a.

Conclusion
----------
I was surprised to see no clear increase in the reliability of results with the
new within subject and time point normalisation of the data. The effects of
the normalisation are small and don't generally lead to an increase in
mlog10p-values. Based on these results I don't see a clear preference for any
of the two normalisation types, but given that theoretically the within time
point normalisation makes a lot of sense for the sequential analysis, I'll
probably stick with that.