unifying data format for source and meg to use common inference script
Most of the code for inference from MEG channel or source data should be shared
and it was tedious to keep them both synchronised. I realised that I just need
to bring the MEG data into the same format as the source data to be able to use
the same inference script for the sequential analysis. I implemented this in
helpers.py and now only need to make minor changes to infer_sequential.py.