new sequential inference in MEG channels with common inference code
I have run the first inference for MEG channels with infer_sequential.py after
making small changes to the code so that the script is compatible with MEG
channel and source data. In the new version of MNE they also announce changes
in the API for topomap plotting which I now account for.

Results
-------
are stored in
	meg_sequential_201802061518.h5

I have reused the same settings as for the last sequential inference in source
space, i.e., no trial regressors, accev, sum_dot_y and the simple dot
regressors including all dots up to 200 ms before the response in the sequence.
I excluded time outs and normalised the data across trials and the regressors
locally.

The results are very similar to previous analyses. I only had to shift the
location of the peaks by 10 ms in some cases. As in source space, the amplitude
for the permuted data rises towards larger time points (after dot onset), but
my impression is that this is not as strong here. Compared to the last results
plot that I used the average absolute effect sizes are slightly lower now (e.g.
the peak at 170/180 ms was roughly at 0.033 and is now at roughly 0.028). The
topomaps show the same features as before.