sequential analysis for EOG signals
I repeated the sequential analysis that I ran on magnetometers and sources
for the two EOG signals. I decided to high-pass filter the EOG signals with a
cutoff of 2 Hz, before GLM analysis. Otherwise the parameters for the
sequential analysis were the same.

Results
-------
are stored in:
	meg_sequential_201802081546.h5

and further discussed in EOG_email_to_Stefan_20180213. In short, there are
significant group-level effects in the horizontal EOG for dot-x and accev.
Specifically, horizontal EOG is anti-correlated with dot-x from 0 to 200 ms
after dot onset and correlated with dot-x from 200 to 400 ms after dot onset.
Otherwise I did not observe considerable effects. I have no good explanation
for these effects and somehow can't believe that they are eye-movement related
- perhaps a lateralised EEG-potential?