event detection for horizontal EOG signal
Using insights from visual inspection of the high-pass filtered EOG signal I
created an EOG event detection algorithm. This was inspired by the EOG event
detection procedure already existing in MNE, but I was after more subtle events
occurring in the horizontal EOG. The clearest of those events are oscillatory
bursts lasting for about 100 to 150 ms. To detect these I rectified the signal
(absolute values) and smoothed the result whereby I let the smoothing frequency
depend on an expected saccade duration of 120 ms (the saccade duration gives
half the period of the low-pass cutoff frequency of the smoothing filter). I
then use MNE's peak finder to detect the events in the signal. This procedure
also detects consistent deflections from baseline, if they have a fitting
period.

I additionally implemented 3 criteria to exclude unwanted/spurious peaks:

1. peak threshold defined in terms of percentile of data across all time points
   and epochs of one subject
2. distance to edge of epoch window: peaks which are too close to an edge are
   excluded based on the observation that peaks there appear strangely
   frequently and I presume that this is because of edge effects introduced by
   filtering
3. too many peaks in one epoch: Some epochs appeared to be very noisy without
   distinctive events in them. Yet, these epochs tended to exhibit many peaks.
   To exclude these noise peaks from further analysis I exclude peaks from
   epochs which had more than a maximum number of peaks.

I set parameters of these criteria so that I got a reasonable compromise
between false positives and negatives in comparison with quick visual
inspection of a couple of examples (cf. show_events). So the procedure can
probably be improved a lot, but for the effort I invested, the result is
acceptable. My impression is that I get about 50% of visable events with few
false positives, but I anyway don't have a ground truth and don't know what to
look for.

The analysis produces three outputs:

1. eog_events: contains all detected events as snippets of the high-pass
   filtered EOG signal centred around the detected peak
2. trial_times: high-level information about the strongest events in all
   trials / epochs in which at least one event has been detected. Specifically,
   the time of the event and the sign of the average event signal are stored
   together with the RT and response in the corresponding trial. The strongest
   event is defined to be the event with the highest average amplitude within
   the event time window.
3. measures: some statistical measures including
	- N: number of events detected for this subject
	- Nuniq: number of trials with events for this subject
	- r_times: correlation coefficient between time of strongest event
	  and RT
	- r_times_p: the corresponding p-value
	- r_signs: correlation coefficient between average sign of strongest
	  event and choice
	- r_signs_p: the corresponding p-value

Results
-------
are stored in:
	eog_201802081218.h5

and further discussed in EOG_email_to_Stefan_20180213. There is a clear
relationship between the time of the strongest event in a trial and the RT: the
strongest EOG events tend to occur 400 ms after the RT leading to strong
correlations between these event times and the RT. There is no obvious relation
between EOG event signs and choices. Plots demonstrating these effects are
produced with make_eog_figures.py which I will commit later.