source reconstruction of eye/motor vertices for full epoch length
In this reconstruction I retain the source vertices of somatosensory and motor
cortex, paracentral lobular and mid cingulate cortex, premotor cortex and
superior parietal cortex. The idea was to capture cortical areas that could
have something to do with generating eye movements. Because eye movements could
occur before occurence of a dot, I did not use baseline correction and extended
the analysed time period to the full available epoch.

fix: As the covariance for source reconstruction needs to be computed
independent of whether baseline correction is done, I hardcoded the covariance
estimation period to be within the previous baseline period of -0.3 to 0 s from
first dot onset.

I initially underestimated the amount of data that would be generated from this
analysis (the final compressed HDF-file is 69 GB large). Without an explicit
call to collect garbage after/before each subject even computing the std across
subjects at the end would quickly fill up all available memory. This is because
there are 1771 vertices in the selected region for each of 34 subjects,
480 trials and 280 time points meaning that the data of one subject is already
about 2 GB including the index. Additionally, loading of the data from the HDF
store for one subject uses a lot of temporary memory that is not automatically
garbage collected afterwards meaning that loading a single subject can make
memory consumption jump from 34% to 79% even though only 42% are used after
call to gc.collect() after loading.

Source reconstruction results are stored in:
	source_epochs_allsubs_HCPMMP1_5_8_201802081827.h5