attempt to find the sources of EOG events
I assumed that all EOG events are generated by the same brain process (ideally:
eye movement generation) and tried to find the location of this process by
looking for sources in the brain which modulate their amplitude around an EOG
event. Other than in previous analyses I could not predict whether a source
current should increase or decrease with respect to the baseline, hence my
concentration only on fluctuations of amplitude, i.e., the absolute value of
the estimated source currents. I initially thought about using a special
generalised linear model to account for the positivity of the data, but the
simplest and standard procedure for this kind of scale data is to simply
log-transform the data and run a standard general linear model in log-space.

As regressors I only use an intercept and an EOG event regressor that is 1 in a
desired time window around an EOG event and is otherwise 0. This means that my
analysis looks for source vertices in which the amplitude of the signal either
increases or decreases on average across the time window around EOG events and
does so consistently across trials. Inspired by the button press motor cortex
activations I set this time window to be -100 to 50 ms from EOG peak event time
(beyond 50 ms after the button press the source response inverted).  Because
the GLM in this analysis is over all time points (concatenated across trials),
there is only one estimated parameter per source vertex.

Results
-------
are stored in
	eog_source_201802091232.h5

and further discussed in EOG_email_to_Stefan_20180213. I did not find any
significant effects after mutliple comparison correction. However, it is odd
that _all_ vertices exhibit a positive effect of the EOG event regressor, i.e.,
around EOG events the amplitude of the signal rose in all considered source
vertices compared to the baseline of an amplitude of 1. In the permuted runs,
on the other hand, parameter estimates were mostly negative, but also positive,
i.e., this was not an effect of the log-transform. Given that most EOG events
occur late in the trial around 400 ms after the response, this result could
point to an overall increase in signal amptlitudes after the response. This
makes sense for the motor areas, although I guess that 400 ms after the
response is too late for motor effects to be seen, but why does the increase
also occur in superior parietal areas (although less so)? Also, as I only
permuted across trials, but not time points, the signal amplitudes are not
generally increasing throughout a trial, but the increase must be linked to the
specific trials, i.e., the response which varied across trials.

When setting the colormap so that the largest effects can be singled out, I
observe the largest clusters of large high t-values in supplementary motor and
mid cingulate areas, but also in (left) pre-motor areas. The cluster with the
highest t-value, however, is around the very lateral end of area 3a in the left
hemisphere without any effect in the corresponding right area. I don't know how
to interpret this, though.