investigating evoked (average) signal around 400 ms after response for EOG
My analysis of EOG events suggested that they occur mostly around 400 ms
after the response. At the same time the investigation of the (amplitude of
the) sources of all EOG events was not very insightful. Consequently, I hoped
to better identify the sources of EOG events by looking at the average source
signals occurring with the major EOG event. In principle I should have got the
same results for the intercept in a response-aligned singledot analysis, but I
hadn't done that, yet (I think today), and it was easy to put together a new
script for simple averaging. Probably I should consolidate the analysis within
a singledot analysis at some point, although that's not important to me now.

Results
-------
are stored in
	evoked_source_201802091710.h5

and are further discussed in EOG_email_to_Stefan_20180213. The surprising
result is that the average signals from about 300 to 440 ms after a response
exhibit a pattern that is extremely similar to the (inverted) pattern of
activations that I observed for the dot x-correlations within about 400 to 500
ms after dot onset. Across this time period of 300 to 440 ms after response I
observe the most consistent activation at the border between areas 6a and 6d,
but only in the left hemisphere. The right hemisphere has similar patterns, but
they are much weaker. Thus far I have no explanation for why I observe these
pattern similarities for two seemingly completely different analyses.