extend time window of response-aligned evoked signal
In the attempt to further investigate the reason for the evoked signal pattern
I computed the full evoked signal -1000 to 1000 ms after the response. I had
before checked on a single vertex that this is roughly the time window in which
interesting things happen. Also, outside this time window the evoked estimates
become quite noisy due to small number of data points.

In my previous analyses I never really looked at the evoked signal, but only at
the response signal which differentiates between left and right responses. It
appears now that there is an asymmetry between signal amplitudes for left and
right responses in the contralateral motor areas which let the average/evoked
signal represent the same process, i.e., the evoked signal roughly exhibits the
brain patterns that I had observed for the response regressor before except
that the effects appear a bit stronger in the evoked signal and also there are
some extra effects in the evoked signal, too.

I had never looked at consistent brain signals beyond 300 ms of the response,
so didn't notice that these are similar to the motor signals 20 ms after the
response. When directly comparing the evoked patterns at 20 and 380 ms, it
appears that the pattern at 380 ms is just an amplitude-reduced version of the
pattern at 20 ms, but the pattern is inverted and it is only strong in the left
hemisphere. Hame has sent me a review of sensorimotor rhythms
(http://dx.doi.org/10.1016/j.expneurol.2012.08.030) which suggests that there
is a motor rebound in the MEG signal beginning around 400 ms after response.
Although this may suggest that the 300-500 ms evoked signal is some motor
after-effect which may be unrelated to the 400 ms EOG events, it's hard to see
the connection between the oscillatory findings in the beta band in the review
and our evoked signals.