new sequential analysis on motor and parietal sources
For the comparison of spatial patterns related to evoked signals at response
time and representations of dot_x I needed to update the results for the
sequential dot_x analysis. This was because I compute evoked responses on the
extended data set of vertices across premotor, motor, paracentral and superior
parietal areas, but only had sequential analysis results for premotor and motor
areas. Also, the previous sequential analysis on motor vertices did not use the
updated sequential analysis code including local normalisation and consistent
exclusion of trials based on timed out response or closeness to response time.
Specifically, the sequential analysis only included a fixed number of dots
instead of now including as many dots as available up to the desired time
before the response.

Because the new vertices data set is huge, I had to take care of memory usage
during the analysis - I included calls to gc.collect after loading subject data
to prevent the memory from clogging up with useless stuff from previous loads.
The local normalisation based on mean and std also uses quite a bit of memory,
but pandas clears that after the computations automatically. To run this
analysis I need at least 50% of my 16GB of memory available, but mostly only
during the load process from the HDF.

Results
-------
are in:
	source_sequential_201803131115.h5

The effects are reasonably big (in terms of t-values) and nicely show that the
spatial patterns resulting from the two analyses are comparable, even though
the x-correlations are a bit weaker.