weeding out numerical issues when computing DM condition numbers
Testing new infer_singledot.py I stumbled upon an error, because one regressor
(response) became constant across trials for very few number of trials for one
subject. The issue was that my local normalisation of the design matrix still
tried to normalise the constant regressor leading to NaN in the design matrix
for that regressor which in turn threw an exception in eigenvalue decomposition
during computation of the condition number. So I now identify all constant
regressors (not only intercept) during normalisation and do not normalise them.

Additionally I came across the issue that I got complex condition numbers for
design matrices which were not full rank. I believe that this was because the
eigenvalues that should have been 0 were actually negative (very close to 0,
but negative, e.g., -1e-17). To resolve this I now map all negative eigenvalues
to 0 before computing the condition number.