export of significant results to csv, switch to lateral views
I now generate csv files containing all significant clusters and the values
shown in the brain figure. This is to easily copy them into supplementary
tables, although I could also just attach them as csv-files.

Because the significant effects in 300-500 ms after dot onset are not in
occipital areas, I now use the lateral views to show the results which make it
easier to see effects in motor cortex.