split response-aligned analysis in inference and figures + average delays
The response aligned analysis lacked statistical power, because for each time
point from response I only could use one dot which was tied to the time point
through a delay, i.e., for each time point I only had maximally number of
trials data points while I had a multiple of that for the sequential analysis.
This meant that almost no correlations were significant after multiple
comparison correction for time points earlier than 200 ms before the response
although the population mean indicated an effect in comparison to
y-coordinates.

Originally, I tried to overcome this by using GPs to take the temporal
correlations into account which should remove some need to correct for multiple
comparisons across time points. However, while writing the paper, I realised
again that the GP results can depend quite strongly on the time window chosen.
Also I stumbeld across my approach to divide the variance across subjects by
the number of subjects to get the data variance for the heteroscedastic GP. The
results appeared to support this approach, but I couldn't remember the
reasoning and it felt wrong. I now remember why I did it: The GP is not
supposed to model the subject-specific data directly, but it models the
underlying mean function. I'm observing the mean across subjects and the
variance of the mean is the variance across subjects divided by the number of
subjects. So my approach was correct and results were ok, but that still left
the issue that results could vary considerably depending on which time points
were included in the analysis.

So I looked for an alternative to increase the power of the analysis and
realised that I could just re-estimate regression coefficients for different
dot delays and then average coefficients across delays. The underlying
assumption is that results with different delays are approximately independent.
I first questioned whether this would be the case: For any given time point t0
changing the delay will only change the association between data and dot
x-coordinates every 100 ms. For example, in a trial with response at 840 ms t0
= -230 is associated with the 4th dot for a delay of 300 ms. For a delay of 310
ms t0 is still associated with the 4th dot. For delay=320 this switches to the
3rd dot, but stays with the 3rd dot until delay=420. So there are always 10
settings of delay for which the association between data and regressor are the
same. However, across trials a single step in delay can lead to a sizable
number of changes in associations, because each trial has its own response time
and therefore regressor associations will change differently across trials for
the same delay. This should induce enough variation in the data / model so that
all delays should be informative on their own.

As the inference for all delays within 300 to 500 ms takes a while, I decided
to separate inference from results analysis and plots. Hence, the new
make_respaligned_figures.py. Note that the second-level results generated by
infer_respaligned_source.py are computed within delays. This is not what I
wanted to do. My idea was to average across delays within subjects so that the
second-level results still reflect the across-subject variance of the effects.
This approach means that the statistical tests (across subjects) are somewhat
robust against violations of the indepdence assumption for the delays, because
the within-subject coefficients may become arbitrarily precise without changing
across-subject variance much that determines whether the second-level t-test
will become significant. So I re-compute the second-level results after
averaging across delays in make_respaligned_figures.py.