fix: exception in computation of condition number when only one row in DM
When there is only one row in a design matrix and I normalise that design
matrix, the values will become nan, because pandas returns nan for the
standard deviations. This, in turn, led to exceptions when computing the
condition number of the normalised design matrix. I now prevent exceptions from
being thrown by checking whether the number of rows in the design matrix is
smaller than the number of columns. Then, the matrix is anyway rank deficient
such that the condition number is infinity. This also speeds up the
computations in these cases.