focusing the linearity analysis on evidence vs. response comparison
Although I originally started out only thinking about comparing linear and step
models, I do actually consider 4 different models, because a change in the
regressor values changes the statistical model in an equivalent way as taking
the sign of the regressor value inside the model definition. In the end I'm
interested in two things: i) Is the signal gradual, or is it step-like? and
ii) Is the signal representing stimulus information, or the response? By
changing the regressor values I can get the following models:
	1) stimulus, linear
	2) stimulus, step
	3) response, linear
	4) response, step

Models 2 and 3 are somewhat artificual mixtures of the two genuine x-coordinate
and response regressors: Model 2 artificially discretises x-coordinates so that
the regressor only indicates whether the dot was on the left or the right.
Model 3 artificially linearises the responses by assigning them the magnitude
of the x-coordinate displacement.

For the paper the most interesting question was whether the signal represents
rather the stimulus or the response (question ii). So I have repeated the model
inference, now computing the corresponding model comparison both for linear and
step models. As the interpretation of models 2 and 3, however, is not quite
straightforward, Stefan suggested to remove them from the analysis and just
compare models 1 and 4. This is fine by me and lead to very similar results as
comparing stimulus vs. response models. Because we're framing everything in
terms of evidence now, I called this comparison the evidence vs. response
comparison.

Results are stored in:
	source_linearity_201804111716.h5 (delay=400, [-500, 400]ms, area 4)

which includes all the discussed model comparisons, but only from 500 ms
before the response onwards, because it takes a long time to compute
everything and I wanted to save some computation time. The results provide mild
evidence for evidence encoding before 200 ms before the response, but only in
left primary motor cortex. In right primary motor cortex the data is
inconclusive. Note that these results are only for a single delay of 400 ms,
because I didn't dare to implement the inference for several delays at the same
time. Potentially it doesn't take so much longer to infer with more data as the
time is spent within Stan's C-code, but I already run now into memory issues,
because each Stan chain uses considerable amount of memory (~1.1 GB) which gets
combined in the parent process with some overhead, too, it seems. Limited
memory is also the reason why I don't save the cross-validated log-likelihood
values directly instead of only the mean and std of their difference for model
comparison (there is one value per data point per sample from the posterior per
model).

If it is needed, I could run inference on an expanded data set with several
delays for only one time point, but with only one chain at a time.