moved computation of response-aligned index to common function
To make it easier to later combine design matrices of different delays I wanted
to move the computation of the response-aligned index into the design matrix
generation function. I also didn't like that I just assinged the index of the
data to the DM and only implicitly excluded timed out trials during data
loading. So I now moved all of this in the data loading and DM generation
functions and created a new function 'to_resp_aligned_index' which transforms
a given subject-trial-time index with first dot onset aligned time to one with
response-aligned time.

In tests I also found that my criterion for finding response times that are
multiples of 5 ms wasn't very good, because it missed some response times that
were very close to being a multiple of 5 so that rounding still lead to equal
consecutive times in the resulting response-aligned index. I fixed that now and
verified that the independently created indeces for data and DM are the same.