expanded data (delays) for evidence vs choice comparison
The model comparison between evidence and choice models I ran wasn't
conclusive. Specifically, the right hemisphere appeared to slightly favour the
choice model. So I thought that I could increase the statistical power of the
analysis by expanding the data considering several delays instead of only one,
exactly as I had done for the other response-aligned analysis. My intuition was
that the small benefit for the choice model in the right hemisphere would go
away on the expanded data set. So I chose 5 delays centred on the previous 400
ms (330, 370, 400, 430, 470) and a single time point with a preference for the
cohice model in right M1 (-350 ms) to check this.

The increased amount of data also lead to increased memory consumption. So I
had to serialise the 4 Stan chains. So far I do not compute diagnostics for
the serialised chains manually. To speed things up and save memory I now also
have an option for skipping computation of mixed models and related model
comparisons.

Results
-------
are stored in:
	source_linearity_201804201508.h5
for:
- time = -350
- delays = 330, 370, 400, 430, 470
- nchains = 4
- skip_mix = True

Other than expected, the cross-validated log-likelihood differences did not
change much with the expanded data. Specifically, at -350 ms the previous
analysis with delay=400 indicated a tiny benefit for the evidence model in the
left hemisphere, but a larger benefit for the choice model in the right
hemisphere. Although with the expanded data the error bars are smaller, the
overall pattern remained. The largest change is in the left hemisphere where
the average difference decreased and is now very close to 0 (change from
0.00013 to 0.000012) whereas the right hemisphere difference remained roughly
the same (-0.00031 to -0.00027). These results suggest to me: i) The data is
indeed inconclusive about whether evidence or choice model are better early
before the response or ii) The previous results with single delay of 400 ms are
perhaps already reasonably accurate. i) may be a warranted conclusion, but
looking only at a single time point, I'm making merely informed guesses.