separated source figures from make_sensor_figures + new plot
Mostly I just wanted to add a new plot showing the time courses of average
absolute regression coefficients in source space for all regressors. Then I
decided to move all source figures out of make_sensor_figures.py, because the
name of the file didn't fit to the mix of figures contained in it.