bugfix: ensure that there is a trial count for each subject
Previous code has led to errors when there was no data left for a subject so
that the subject was dropped from the index. Then, the indexing from
trial_counts into subjects did not work anymore. It's also more consistent to
store that there was no data for a subject explicitly.