motor vertices figure for the manuscript
I previously showed results from left and right hemisphere in a simple stitch
of the brain figures, but the results from the right hemisphere could be a bit
confusing, because the colours are swapped for the two types of results
(showing two different things). I now stitch the brain figures together in
python to be able to easily add labels.