first dot onset aligned evidence vs response model comparison in area 4
My idea was that the first dot onset aligned data should exhibit a clearer
evidence effect than the response-aligned data. This also appears to be true
according to the results, but the results are not that much more informative.
Especially in right primary motor cortex the model comparison results are still
inconclusive, but at least the t-test over the mean loodiffs across time now
also indicates a perference towards the evidence model.

Results are in:
	source_linearity_201805280909.h5