bugfix: using correct parcellation + plotting outlines
Burkhard suggested to indicate the 5 areas per hemisphere mentioned in the
caption (top t-value magnitudes) by drawing outlines around them.

Doing this I noticed that outlines and areas didn't quite overlap. This was
because I still used HCPMMP1 to plot the areas. Results are now based on
HCPMMP1_5_8 so I changed this.