bugfix: prevent error when no time outs + follow pandas reindex suggestion
I installed the analysis code on a new computer. The new version of pandas (or
perhaps the shift to Windows) introduced an error for selecting data that
doesn't exist, specifically when I select timed out trials. I now explicitly
check whether there are any timed out trials before trying to access them.

Also, pandas now complains about using .loc[array] when array contains
elements that do not exist in the index. This happens when I select the data
corresponding to the times of the design matrix, because I construct the design
matrix times so that late dots will be mapped to non-existant data times. I
just had to use reindex instead of loc to get rid of the warning and comply
with future behaviour of pandas.