new source reconstructions with loose orientations and sign flips
An eLife reviewer complained that our results follow the anatomical landscape
of the brain (plots showing high-res motor vertices). I have verified that this
is the case: e.g. x-correlations around the central gyrus are positive on one
side of the gyrus and negative on the other. This is not an issue, though, but
is expected for fixed orientation constraints (cf. https://mne-tools.github.io/stable/auto_tutorials/plot_dipole_orientations.html).
I explain it like that: The limited resolution of MEG means that you can't
resolve the dipole flip from one side of a gyrus to another. Instead the
pattern of magnetic fields only suggests a larger overall current flow in a
particular direction. This will manifest itself as a positive current flow on
one side of the gyrus and a negative current flow on the other side. So the
anatomical pattern of effects can be explained by the specific analysis
setting, but the reviewer is right that this is not very informative, or rather
misleading, because it suggests a too high spatial resolution. Also, it makes
average effects within brain areas depend on whether the area covers opposing
sides of a gyrus. So I decided to re-do the source reconstruction accounting
for this.

Loose orientations and current magnitudes
-----------------------------------------
Additionally, Burkhard argued pre-submission that it would be better to loosen
the constraints on dipole orientations during source reconstruction in our
case to better account for inter-individual differences in the brains of our
participants. I now followed the MNE standard settings with loose=0.2 which
weakens the cortex orientation constraints. My first try was then to only store
the magnitude of the source vectors for further analysis (pick_ori=None).
Results of this source reconstruction can be found in:

	source_epochs_allsubs_HCPMMP1_5_8_201807021743.h5 (areas)
	source_epochs_allsubs_HCPMMP1_5_8_201807022008.h5 (motor vertices)

It turned out that these reconstructions are useless for my analysis, because
a) I still did not account for the anatomically dependent sign of the results
during averaging within area, but more importantly b) only using the magnitude
of the source currents destroys all statistical effects. This is because the
large noise in the MEG signal leads to switching signs of the source currents
across different trials such that estimated source currents tend to centre
around 0 with some small deviations on average. When I take the absolute value
of the source currents to get their magnitude, the average magnitude tends to
only follow the variance of the estimated source currents, but not their mean.
My statistical analyses try to estimate average effects, but these will be
masked by variance effects in these data. Consequently, I do not see any
significant effects of dot_x in the regression analysis. Instead there are only
weak effects of sum_dot_y_prev, indicating that the magnitude of source
currents tended to be high in some areas whenever the dots tended to occur in
one vertical half of the screen in one trial. Results of this statistical
analysis are in:

	source_sequential_201807031144.h5

Loose orientations, flips and normal currents
---------------------------------------------
I now am also running a source reconstruction with loose orientation
constraints (loose=0.2), but extracting the source current orthogonal to the
cortical surface instead of the unconstrained magnitude of the source currents.
This can be achieved with pick_ori='normal'. It is not clear in the MNE
documentation whether this option will also only return magnitudes, but I
confirmed that it returns signed source currents.

In extract_label_time_course MNE actually provides an option to aggregate
source currents while accounting for the orientation of the cortex
(mode='mean_flip'). I originally used this option, but discarded it after
looking into retinotopic effects in primary visual cortex (cf. commit
b6c44b59). Stefan and I discarded it at the time based on the misunderstanding
that using fixed orientation constraints already properly aligns the signs of
the source currents so that further flipping becomes unncessary. My recent
considerations detailed here do not support this. Therefore, I re-introduced
mode='mean_flip' in the newest source reconstruction. This should lead to more
areas with significant effects, but it's not entirely clear to me how
situations are handled in which one gyrus connects perpendicular to another
within one area. Ideally I would not aggregate across sources within one area
prior to statistical analysis, but instead run statistical analysis on
individual sources and then aggregate ignoring the sign. This would be
extremely computationally heavy, though, not only because the many sources, but
also because of statistically aggregating magnitudes.

Anyway, results of source reconstruction with loose=0.2, pick_ori='normal'  and
label_mode='mean_flip' are in:

	source_epochs_allsubs_HCPMMP1_5_8_201807041847.h5

Bugfix
------
Although I defined options 'pick_ori' and 'label_mode' at the beginning of the
source reconstruction script, I had hard coded the corresponding arguments to
MNE functions so that the chosen options had no effect.